refactor: regenerate tenant-management proxies
