+++ npm/ng-packs/apps/dev-app/src/environments/environment.ts
@@ -28,11 +28,11 @@ export const environment = {
     },
     AbpPermissionManagement: {
       url: 'https://localhost:44305',
-      rootNamespace: 'Volo.Abp',
+      rootNamespace: 'Volo.Abp.PermissionManagement',
     },
     AbpTenantManagement: {
       url: 'https://localhost:44305',
-      rootNamespace: 'Volo.Abp',
+      rootNamespace: 'Volo.Abp.TenantManagement',
     },
     AbpIdentity: {
       url: 'https://localhost:44305',
