+++ npm/ng-packs/packages/tenant-management/src/lib/models/tenant-management.ts
@@ -1,5 +1,5 @@
 import { ABP, PagedResultDto } from '@abp/ng.core';
-import { TenantDto } from '../proxy/tenant-management/models';
+import { TenantDto } from '../proxy/models';
 
 export namespace TenantManagement {
   export interface State {
