+++ npm/ng-packs/packages/tenant-management/src/public-api.ts
@@ -5,4 +5,4 @@ export * from './lib/enums';
 export * from './lib/models';
 export * from './lib/services';
 export * from './lib/states';
-export * from './lib/proxy/tenant-management';
+export * from './lib/proxy';
